% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciMPN.R
\name{ciMPN}
\alias{ciMPN}
\title{Standard Error / Confidence Interval function}
\usage{
ciMPN(mu, y, n, v, alpha = 0.05, output = FALSE)
}
\arguments{
\item{mu}{getMPN output}

\item{alpha}{alpha value (usually 0.05 for 95 CI)}

\item{output}{if FALSE returns the standard error, if TRUE returns the confidence interval}

\item{dataMPN}{dataframe (or list) named "dataMPN": dataMPN<-data.frame(x=..., n=..., v=...)}
}
\value{
the standard error/confidence interval
}
\description{
Calculate the Standard Error (SE) or Confidence Interval (LL, UL)
}
