% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPN.R
\name{getMPN}
\alias{getMPN}
\title{Most Probable Number (MPN) function}
\usage{
getMPN(y, n, v)
}
\arguments{
\item{n}{A numeric vector containing the number of MPN replicates}

\item{v}{A numeric vector containing the sample volumes}

\item{dataMPN}{dataframe (or list) named "dataMPN": dataMPN<-data.frame(x=..., n=..., v=...)}

\item{x}{A numeric vector containing values of the positive wells.}
}
\value{
The MPN value
}
\description{
Calculates the MPN according to a data.frame (or list).
}
