% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMPN.R
\name{probMPN}
\alias{probMPN}
\title{Probability function}
\usage{
probMPN(mu, y, n, v)
}
\arguments{
\item{mu}{getMPN output}

\item{dataMPN}{dataframe (or list) named "dataMPN": dataMPN<-data.frame(x=..., n=..., v=...)}
}
\value{
the probability over 1
}
\description{
Calculate the probability to obtain a specific MPN value
}
