% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDIL.R
\name{repDIL}
\alias{repDIL}
\title{Dilution vector}
\usage{
repDIL(x, y, z)
}
\arguments{
\item{x}{A numeric value as starting value}

\item{y}{A numeric value as vector length}

\item{z}{A numeric value as division factor}
}
\value{
A vector
}
\description{
Creates a vector with a starting value of x and length of y and dilution of z each incrementation.
}
